Bundle and minify app JavaScript code
Closes #23

The existing page loads 36 JavaScript source files. 25 of these files
are Angular app code split into controller, service, and directive
modules.

This PR bundles and minifies those files to prevent loading 24
additional scripts. Although the scripts do not make up a major portion
of the downloaded bandwidth, they account for almost 75% of the network
requests performed when loading the site.

Now, a single `bundle.js` file is loaded, leaving a total of 12, or one
third of the previous number of scripts. The new size of downloaded code
drops from about 168 KB of Angular app code to 55 KB.

Further, a 299 KB source map is built and available for someone who
opens the browser dev tools. It provides a mapping back to the original
un-minified source code but will only be downloaded when the dev tools
are open and inspecting.

**There should be no visual or functional changes in this PR**

**Further work before merging**

 - [ ] Create `package.json` with appropriate dependencies and scripts
 - [ ] Update project README with any notes required for the new build
   step we have introduced (bundle and minify with webpack)
 - [ ] Determine if it's possible to automatically bundle and minify
   JavaScript code without manual intervention - maybe with something
   like [flask-webpack](https://github.com/nickjj/flask-webpack)

**Notes**

 - I have tried to keep this PR as small as possible, meaning I have not
   attempted to alter anything beyond incorporating **webpack** and
   bundling the app-specific code.

 - We have a few more project-specific JavaScript files that we can
   add to the bundle.

 - If we want to we can bundle the additional libraries to further
   reduce the number of network requests and benefit from futher
   minification. This brings a couple of implications:

    - The vendor libraries become intrinscically version-pinned to
      whatever we choose. Even if something were to happen to the
      external libraries, we would still have control over the code
      we build in our app.

    - We would likely experience much higher compression as webpack
      tries to eliminate unreachable or unused code paths.

	- Unfortunately, if we do bundle the vendor libraries, then we lose
	  the possibility of piggy-backing off of the cached requests from
      other sites on the web. I do not see this as a practical situation
      since we're already loading specific (and old) versions of the
      libraries we use.

 - Introducing **webpack** is just the beginning. We can start using
   code transforms with tools like Babel to allow more sane coding
   styles without polluting the production code.